Land classification tool
A land classification tool that allows to intersect DEMs, polygon land use layers and polygon subcatchment layers for big regions. It manages to carry out the intersections subdividing the extent of the region into smaller ones. Includes a tool to export the generated classification into separate subcatchment-elevation layers into a single GPKG or into a folder as Shapefiles.